server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
email-validator==2.0.0
Werkzeug==2.3.7
gunicorn==21.2.0